# packages/api/llm_utils/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chain Swarm API",
    "description": "API for accessing chain swarm",
    "version": "0.2.0"
  },
  "paths": {
...
    "/substrate/{network}/balance-series/address/{address}/history": {
      "get": {
        "tags": [
          "balance-series"
        ],
        "summary": "Get Address Balance History",
        "description": "Retrieves historical balance snapshots for a specific address on a blockchain network.\n\nThis endpoint provides a paginated list of balance snapshots at fixed 4-hour intervals, showing free, reserved, staked, and total balances along with changes between periods. Optionally filter by time range and assets.",
        "operationId": "get_address_balance_history_substrate__network__balance_series_address__address__history_get",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain network identifier",
              "title": "Network"
            },
            "description": "The blockchain network identifier",
            "example": "torus"
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain address to query",
              "title": "Address"
            },
            "description": "The blockchain address to query",
            "example": "5C4n8kb3mno7i8vQmqNgsQbwZozHvPyou8TAfZfZ7msTkS5f"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number for pagination",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number for pagination"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of balance snapshots per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of balance snapshots per page"
          },
          {
            "name": "assets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
              "title": "Assets"
            },
            "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
            "example": [
              "TOR"
            ]
          },
          {
            "name": "start_timestamp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start timestamp in milliseconds (Unix timestamp)",
              "title": "Start Timestamp"
            },
            "description": "Start timestamp in milliseconds (Unix timestamp)",
            "example": 1742815440000
          },
          {
            "name": "end_timestamp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End timestamp in milliseconds (Unix timestamp)",
              "title": "End Timestamp"
            },
            "description": "End timestamp in milliseconds (Unix timestamp)",
            "example": 1842815440000
          }
        ],
        "responses": {
          "200": {
            "description": "Balance history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Address not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/substrate/{network}/balance-series/address/{address}/current": {
      "get": {
        "tags": [
          "balance-series"
        ],
        "summary": "Get Current Address Balance",
        "description": "Retrieves the latest balance snapshot for a specific address on a blockchain network.\n\nThis endpoint returns the most recent balance data including free, reserved, staked, and total balances for the specified address and assets.",
        "operationId": "get_address_current_balance_substrate__network__balance_series_address__address__current_get",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain network identifier",
              "title": "Network"
            },
            "description": "The blockchain network identifier",
            "example": "torus"
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain address to query",
              "title": "Address"
            },
            "description": "The blockchain address to query",
            "example": "5C4n8kb3mno7i8vQmqNgsQbwZozHvPyou8TAfZfZ7msTkS5f"
          },
          {
            "name": "assets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
              "title": "Assets"
            },
            "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
            "example": [
              "TOR"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Current balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Address not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/substrate/{network}/balance-series/address/{address}/changes": {
      "get": {
        "tags": [
          "balance-series"
        ],
        "summary": "Get Address Balance Changes",
        "description": "Retrieves balance changes analysis for a specific address on a blockchain network.\n\nThis endpoint provides a paginated list of balance changes, showing periods where the balance changed along with the magnitude of changes. Results are ordered by change magnitude (largest changes first).",
        "operationId": "get_address_balance_changes_substrate__network__balance_series_address__address__changes_get",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain network identifier",
              "title": "Network"
            },
            "description": "The blockchain network identifier",
            "example": "torus"
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain address to query",
              "title": "Address"
            },
            "description": "The blockchain address to query",
            "example": "5C4n8kb3mno7i8vQmqNgsQbwZozHvPyou8TAfZfZ7msTkS5f"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number for pagination",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number for pagination"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of balance changes per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of balance changes per page"
          },
          {
            "name": "assets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
              "title": "Assets"
            },
            "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
            "example": [
              "TOR"
            ]
          },
          {
            "name": "min_change_threshold",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum absolute change threshold to filter by (in token units)",
              "title": "Min Change Threshold"
            },
            "description": "Minimum absolute change threshold to filter by (in token units)",
            "example": 1000.0
          }
        ],
        "responses": {
          "200": {
            "description": "Balance changes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Address not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/substrate/{network}/balance-series/aggregations/{period}": {
      "get": {
        "tags": [
          "balance-series"
        ],
        "summary": "Get Balance Aggregations",
        "description": "Retrieves balance aggregations for daily, weekly, or monthly periods.\n\nThis endpoint provides aggregated balance data over different time periods, showing end-of-period balances and cumulative changes. Useful for trend analysis and reporting across multiple addresses and assets.",
        "operationId": "get_balance_aggregations_substrate__network__balance_series_aggregations__period__get",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain network identifier",
              "title": "Network"
            },
            "description": "The blockchain network identifier",
            "example": "torus"
          },
          {
            "name": "period",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(daily|weekly|monthly)$",
              "description": "Aggregation period",
              "title": "Period"
            },
            "description": "Aggregation period",
            "example": "daily"
          },
          {
            "name": "addresses",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of addresses to filter by. If not provided, returns data for all addresses.",
              "title": "Addresses"
            },
            "description": "List of addresses to filter by. If not provided, returns data for all addresses.",
            "example": [
              "5C4n8kb3mno7i8vQmqNgsQbwZozHvPyou8TAfZfZ7msTkS5f"
            ]
          },
          {
            "name": "assets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
              "title": "Assets"
            },
            "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
            "example": [
              "TOR"
            ]
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date in YYYY-MM-DD format",
              "title": "Start Date"
            },
            "description": "Start date in YYYY-MM-DD format",
            "example": "2025-01-01"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date in YYYY-MM-DD format",
              "title": "End Date"
            },
            "description": "End date in YYYY-MM-DD format",
            "example": "2026-01-31"
          }
        ],
        "responses": {
          "200": {
            "description": "Balance aggregations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "400": {
            "description": "Invalid period specified"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/substrate/{network}/balance-series/volume-series": {
      "get": {
        "tags": [
          "balance-series"
        ],
        "summary": "Get Balance Volume Series",
        "description": "Retrieves network-wide balance activity volume metrics over time.\n\nThis endpoint provides aggregated balance activity data showing the volume of balance changes across all addresses for each time period. It includes metrics like active addresses count, total balance change volumes, and breakdowns by balance type (free, reserved, staked). Results are grouped by time periods and assets, useful for analyzing network-wide balance activity trends.",
        "operationId": "get_balance_volume_series_substrate__network__balance_series_volume_series_get",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain network identifier",
              "title": "Network"
            },
            "description": "The blockchain network identifier",
            "example": "torus"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number for pagination",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number for pagination"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of volume series records per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of volume series records per page"
          },
          {
            "name": "assets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
              "title": "Assets"
            },
            "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
            "example": [
              "TOR"
            ]
          },
          {
            "name": "start_timestamp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start timestamp in milliseconds (Unix timestamp)",
              "title": "Start Timestamp"
            },
            "description": "Start timestamp in milliseconds (Unix timestamp)",
            "example": 1742815440000
          },
          {
            "name": "end_timestamp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End timestamp in milliseconds (Unix timestamp)",
              "title": "End Timestamp"
            },
            "description": "End timestamp in milliseconds (Unix timestamp)",
            "example": 1842815440000
          }
        ],
        "responses": {
          "200": {
            "description": "Balance volume series retrieved successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}